requirements.txt
python_docx
openpyxl
